[Dev] Invocation result trait decomposition.
